func (expr *CXExpression) RemoveOutput() { 	if len(expr.Outputs) > 0 { 		expr.Outputs = expr.Outputs[:len(expr.Outputs)-1] 	} } CXExpression.go: Collect the definition of CXExpression and all methods on it.
